chore: create separate cytometry graph header component
