+++ src/components/ViewController/CellMasksLayerSection/GraphFilters/CytometryGraph/CytometryHeader/CytometryHeader.types.ts
@@ -1,0 +1,3 @@
+export type CytometryHeaderProps = {
+  availableProteinNames: string[];
+};
