+++ src/components/ViewController/CellMasksLayerSection/GraphFilters/CytometryGraph/CytometryHeader/index.ts
@@ -1,0 +1,2 @@
+export * from './CytometryHeader';
+export * from './CytometryHeader.types';
